Refactor testing; cover PartitionAwarePrimaryKey too
